Fixed issue where PiP streams would load as null
Electron asar archives are read-only, so fs write ops were failing in
fileOps.js. Refactored to avoid using fs. Info needed to show the right
PiP stream is now stored in a react component, PipMarkup.js, and it is
rendered to the dom at app startup using pipContainer.html.

Also added a preloader window for when the app is starting up